feat: デザインの微調整 (#28)
* improve: サイドバーのパディングが左右で異なっていた問題を修正

* feat: ドロップダウン内のサポートサーバーやドキュメントのリンクを削除

* feat: ホームページのデザインを変更

* Merge remote-tracking branch 'origin/main' into improve/design
